Do not hydrate media or IGTV loaded with profile data (#362)
* Mark IMAP client test as skipped when IMAP extension is not loaded

This prevents having the test suite fail on platforms when the extension
is not loaded.  The test is marked as skipped so the developer knows
that the full test suite was not run

* Add a convenience script to run unit tests with "composer test"

* Do not hydrate media or IGTV loaded with profile data

Instagram no longer reliably returns medias or IGTVs with the call to
get profile information.  Instead, the user must explicitly make a call
to `\Instagram\Api::getMoreMedias` in order to fetch medias or to
`getMoreIgtvs` to fetch IGTVs